+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2026-02-09
+- feat: Add a WordPress snippet to trigger Netlify builds when posts are saved, published, or deleted.
+
 ## 2026-02-06
 - feat: Add a mobile slide-out menu for easier navigation.
 - feat: Add category-based archive navigation and a Pages menu.